+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Removed dependencies on `lodash` and `fs-extra`.
+
 ## [3.0.0] - 2021-04-03
 
 ### Changed